Extract shooting tactics into seperate classes
