+++ Model/Gunner.cs
@@ -21,10 +21,11 @@ namespace Vsite.Oom.Battleship.Model
             shipsToShoot = new List<int>(shipLengths.OrderByDescending(l => l));
             ShootingTactics = ShootingTactics.Random;
             squareTerminator = new SquareTerminator(rows, columns);
+            targetSelect = new RandomShooting(evidenceGrid);
         }
         public Square NextTarget()
         {
-            lastTarget = SelectTarget();
+            lastTarget = targetSelect.NextTarget(shipsToShoot[0]);
             return lastTarget;
         }
         public void ProcessHitResult(HitResult hitResult)
@@ -35,7 +36,7 @@ namespace Vsite.Oom.Battleship.Model
             squaresHit.Add(lastTarget);
             if (hitResult == HitResult.Sunken)
             {
-                squaresHit.Add(lastTarget);
+
                 var toEliminate = squareTerminator.ToEliminate(squaresHit);
                 foreach (var sq in toEliminate)
                     evidenceGrid.MarkHitResult(sq, HitResult.Missed);
@@ -54,6 +55,7 @@ namespace Vsite.Oom.Battleship.Model
             if (hitResult == HitResult.Sunken)
             {
                 ShootingTactics = ShootingTactics.Random;
+                targetSelect = new RandomShooting(evidenceGrid);
                 return;
             }
             if (hitResult == HitResult.Hit)
@@ -62,9 +64,11 @@ namespace Vsite.Oom.Battleship.Model
                 {
                     case ShootingTactics.Random:
                         ShootingTactics = ShootingTactics.Surrounding;
+                        targetSelect = new SurroundShooting(evidenceGrid, squaresHit);
                         return;
                     case ShootingTactics.Surrounding:
                         ShootingTactics = ShootingTactics.Inline;
+                        targetSelect = new InlineShooting(evidenceGrid, squaresHit);
                         return;
                     case ShootingTactics.Inline:
                         return;
@@ -73,75 +77,6 @@ namespace Vsite.Oom.Battleship.Model
 
         }
 
-        private Square SelectTarget()
-        {
-            switch (ShootingTactics)
-            {
-                case ShootingTactics.Random:
-                    return SelectRandomly();
-                case ShootingTactics.Surrounding:
-                    return SelectFromArround();
-                case ShootingTactics.Inline:
-                    return SelectInline();
-                default:
-                    Debug.Assert(false);
-                    return null;
-            }
-        }
-
-        private Square SelectRandomly()
-        {
-            var placements = evidenceGrid.GetAvailablePlacements(shipsToShoot[0]);
-            // create simple array of sqaures from arrays of arrays
-            var allCandidates = placements.SelectMany(seq => seq);
-            // create groups with individual squares
-            var groups = allCandidates.GroupBy(sq => sq);
-            // find the number of squares in largest group
-            var maxCount = groups.Max(g => g.Count());
-            // filter only froups that have maxCount elements
-            var largestGroups = groups.Where(g => g.Count() == maxCount);
-            // fetch keys from each group (i.e. square that represents the group)
-            var mostCommon = largestGroups.Select(g => g.Key);
-            if (mostCommon.Count() == 1)
-                return mostCommon.First();
-            int index = random.Next(0, mostCommon.Count());
-            return mostCommon.ElementAt(index);
-
-        }
-        private Square SelectFromArround()
-        {
-            List<IEnumerable<Square>> arround = new List<IEnumerable<Square>>();
-            foreach (Direction direction in Enum.GetValues(typeof(Direction)))
-            {
-                var l = evidenceGrid.GetSquaresNextTo(squaresHit.First(), direction);
-                if (l.Count() > 0)
-                    arround.Add(l);
-            }
-            if (arround.Count == 1)
-                return arround[0].First();
-            var ordered = arround.OrderByDescending(ls => ls.Count());
-            int maxLen = ordered.First().Count();
-            if (maxLen > shipsToShoot[0] - 1)
-                maxLen = shipsToShoot[0] - 1;
-            var longest = ordered.Where(ls => ls.Count() >= maxLen);
-            int index = random.Next(0, longest.Count());
-            return longest.ElementAt(index).First();
-        }
-        private Square SelectInline()
-        {
-            var l = evidenceGrid.GetSquaresInline(squaresHit);
-            if (l.Count() == 1)
-                return l.ElementAt(0).First();
-            var ordered = l.OrderByDescending(ls => ls.Count());
-            int maxLen = ordered.First().Count();
-            if (maxLen > shipsToShoot[0] - squaresHit.Length)
-                maxLen = shipsToShoot[0] - squaresHit.Length;
-            var longest = ordered.Where(ls => ls.Count() >= maxLen);
-            int index = random.Next(0, longest.Count());
-            return longest.ElementAt(index).First();
-        }
-
-
         private Square lastTarget;
         private Grid evidenceGrid;
         private List<int> shipsToShoot;
@@ -149,6 +84,7 @@ namespace Vsite.Oom.Battleship.Model
         public ShootingTactics ShootingTactics { get; private set; }
         private Random random = new Random();
         private ISquareTerminator squareTerminator;
+        private ITargetSelect targetSelect;
 
     }
 }
