+++ Model/ITargetSelect.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Vsite.Oom.Battleship.Model
+{
+    public interface ITargetSelect
+    {
+        Square NextTarget(int shipLength);
+    }
+}
