+++ Model/RandomShooting.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Vsite.Oom.Battleship.Model
+{
+    public class RandomShooting : ITargetSelect
+    {
+        public RandomShooting(Grid evidenceGrid)
+        {
+            this.evidenceGrid = evidenceGrid;
+        }
+        public Square NextTarget(int shipLength)
+        {
+            var placements = evidenceGrid.GetAvailablePlacements(shipLength);
+            // create simple array of sqaures from arrays of arrays
+            var allCandidates = placements.SelectMany(seq => seq);
+            // create groups with individual squares
+            var groups = allCandidates.GroupBy(sq => sq);
+            // find the number of squares in largest group
+            var maxCount = groups.Max(g => g.Count());
+            // filter only froups that have maxCount elements
+            var largestGroups = groups.Where(g => g.Count() == maxCount);
+            // fetch keys from each group (i.e. square that represents the group)
+            var mostCommon = largestGroups.Select(g => g.Key);
+            if (mostCommon.Count() == 1)
+                return mostCommon.First();
+            int index = random.Next(0, mostCommon.Count());
+            return mostCommon.ElementAt(index);
+        }
+
+        private readonly Grid evidenceGrid;
+        private Random random = new Random();
+
+    }
+}
